Snap correctly when scroll operation goes outside scroll bounds
In the current snap selection algorithm there is an inherent assumption that the selected snap
position in the other axis is inside the scroller bounds. Breaking this invariant in snap selection
algorithm causes it to not find any valid candidate.

In reality, this variant could have been violated for example the  intended and base position for an
snap operation with intended end and direction could  be outside the bound quite easily e.g., when
for example via Element.scrollBy(-100,-100)  in script of when user flings diagonally in a
horizontal only scroller.

This fix updates the logic so that instead of making the assumption we actively  enforce it by
clamping the out-of-bound values. An alternative approach may have been to perhaps sanitize
the input to snap selection logic but the current approach seems simpler and more robust.

Note that the problem applies to all snapping but  manifested itself most obviously when
|scroll-snap-stop:always| was used because in normal snapping even if fling didn't snap correctly we
would re-snap at the end of the fling so it was less obvious.

Other minor changes:
 - correct spelling error  s/cros/cross/
 - reuse base::ClampToRange to clamp

TEST: wpt/css/css-scroll-snap/scroll-snap-stop-always.html, wpt/css/css-scroll-snap/scrollTo-scrollBy-snaps.html

Bug: 823998
Change-Id: I85ff8b2c87518358805064375e0386606aad7b66